Remove QuantumCircuits.jl (#561)
### Summary

Unfortunately the link to this project is broken, and the contact email
is no longer working so I can't update the link. I'm removing this
project, but the maintainer can always re-apply with updated links in
the future.

I did a quick check, and all other projects have working links.