[BUG] Desactived depreciation test & add test for website builder (#586)
<!--
⚠️ If you do not respect this template, your pull request will be
closed.
⚠️ Your pull request title should be short detailed and understandable
for all.
⚠️ Also, please add a release note file using reno if the change needs
to be
  documented in the release notes.
⚠️ If your pull request fixes an open issue, please link to the issue.

- [ ] I have added the tests to cover my changes.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the CONTRIBUTING document.
-->

### Summary

The test depreciation test can't be used right now until the new version
of qiskit isn't largely used. It come from we still used the old way to
check the qiskit version of projects in order to not get the terra
version. It will be able to revert this change when qiskit terra gonna
totally disappeared.

Cause :

https://github.com/Qiskit/ecosystem/blob/0212a863ffdc1ece3964285ac295db6565500332/ecosystem/templates/configured_tox.ini#L32
As we want to keep the qiskit version and not the terra version we have
to keep it this way for now.


### Details and comments
- [x] Update test `test_tests_runner_on_simple_repo`
- [x] Add test `test_website_builder`

---
Closes #572

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>