Review titles (#614)
### Summary

Project titles are currently a mixture of "normal" titles (e.g. "Quantum
Random Access Optimization"), and slugified titles (e.g.
qiskit-nature-pyscf-dft-embedding). This PR changes many of the titles
to their "normal" titles.

### Details and comments

- I chose "normal" over slugified due to the website styling (IMO
slugified is strange in a non-monospaced font)
- I got these titles by clicking through to the project and using the
title of the README
- I did not change any projects I was unsure about (e.g. the
"pytorch-quantum" README title is "Quantum Computing in PyTorch"; it's
not obvious they'd be happy to stylize it as "PyTorch Quantum")