Remove testing feature (#656)
### Summary

Removes the testing feature of the Qiskit Ecosystem.

This feature would regularly clone projects and run their tests against
different versions of Qiskit to check for regressions and
incompatibility issues. This was a great initiative and an impressive
technical feat.

Last year, we decided to discontinue this feature. Since December last
year, these test results are not publicly available, but the tests are
still running each week. This PR fully disables the tests to save CI
resources, and removes all code related to them.

### Details and comments

During the cleanup I also removed some lingering references to
`members.json` (also no longer needed), and the concept of "tiers",
which we no longer display on the website either.