Revise labels logic (#671)
## Restructure `labels.json`

The JSON file was of the form:

```json
{
  "name": "Label name",
  "description": "Label description"
}
```

but whenever we read the file, we want it in the form

```json
{
  "Label name": "Label description"
}
```

This PR converts the file to the latter form.

## Don't automatically add new labels

On detecting a new label in a repository, the script would add it to
`labels.json`.

I've removed that behaviour as it would silently add new labels
(including misspellings and lower/uppercase duplicates), and not include
a description. We should only add new labels intentionally.