Generalize website generation (#688)
### Summary

Adds the ability to generate HTML that can be embedded into a markdown
document.
You may find it easier to review each commit separately.

- (3ec7a4818f56de487638a245caeb6bd639a430e8) Makes the license field
optional. A license is still a requirement for ecosystem projects, this
change just allows building the website if it isn't set in the database.
- (d24a4ab3720d889db3131a00e22e301a8dc05756) Add 'standalone' option to
generate a page that can be embedded in another page.
This basically means exluding the head/body tags, removing some CSS, and
removing indentation (to play nicely with markdown).
- (92000f26903a3d5bcf05c3f4d08d12e52c67eb1f) Specify output filename
(not just folder) in case you don't want the page to be called
`index.html`.
- (e98f6e5e5a26744e29157142b5ab58d8f7bf9dbf) Add custom css
capabilities. This lets users add custom CSS rules, specified in the
`website.toml` file.