Fix submission process (#701)
### Summary

The submission process was broken as some entries in the issue template
were being ignored by the parser. This PR fixes that and makes the
submission process more robust so similar issues don't occur in the
future.

### Details

Currently, to add a new field to the issue template, you need to edit
four different files to make sure the field makes it through to the
database. This PR avoids that by explicitly mapping field IDs in the
issue template to attributes of the repository model and writing the
entry immediately. After this change, you only need to edit the issue
template and the repository model.

I've also added a test check the fields in the issue template match
fields in the repository model.

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>